requirements.txt
aiohttp==3.6.2
aiohttp-cors==0.7.0
alabaster==0.7.12
argh==0.26.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
bandit==1.6.2
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
dblpy==0.4.0
discord.py==1.5.1
doc8==0.8.0
docutils==0.16
flake8==3.8.4
futures==3.1.1
gitdb==4.0.5
GitPython==3.1.11
goslate==1.5.1
idna==2.8
imagesize==1.2.0
isort==4.3.21
Jinja2==2.10.3
lazy-object-proxy==1.4.3
livereload==2.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.2
packaging==20.1
pathtools==0.1.2
pbr==5.4.4
Pillow==7.2.0
port-for==0.3.1
psutil==5.6.7
pycodestyle==2.6.0
pycparser==2.19
pyflakes==2.2.0
Pygments==2.5.2
pylint==2.6.0
PyNaCl==1.3.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
ratelimiter==1.2.0.post0
requests==2.22.0
restructuredtext-lint==1.3.0
rstcheck==3.3.1
six==1.13.0
smmap==3.0.4
snowballstemmer==2.0.0
soupsieve==1.9.5
Sphinx==3.1.2
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-asyncio==0.2.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-trio==1.1.2
stevedore==1.31.0
toml==0.10.1
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.7
watchdog==0.9.0
websockets==6.0
wrapt==1.11.2
yarl==1.4.2
